requirements.txt
carla == 0.9.14
casadi == 3.6.5
PyYAML == 6.0.1
numpy == 1.21.6
matplotlib == 3.5.3
python-decouple == 3.8
opencv-python == 4.2.0.32
networkx
